Implement login functionality with context and modal, add Dockerfile for development environment, and update HTML and CSS for styling
