+++ dv-marketplace/index.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/dv.xhtml" />
+ 
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Dataverse Marketplace</title>
   </head>
