+++ dv-marketplace/src/components/context/UserContextProvider.tsx
@@ -11,37 +11,45 @@ export type User = {
 
 export type UserContextType = {
     user: User | null;
+    setUser: (user: User | null) => void;
     userRef: MutableRefObject<User | null>;
     login: (user: User) => void;
     logout: () => void;
+    show: boolean;
+    setShow: (show: boolean) => void;
 };
 
 export const UserContext = createContext<UserContextType>({
     user: null,
+    setUser: () => {},
     userRef: { current: null },
     login: () => {},
     logout: () => {},
+    show: false,
+    setShow: () => {},
 });
 
 const UserContextProvider = ({ children }: { children: React.ReactNode }) => {
     
     const [user, setUser] = useState<User | null>(null);
-    const userRef: MutableRefObject<User | null> = useRef(null);
+    const [show, setShow] = useState(false);
 
-    useEffect(() => {
-        userRef.current = user;
-    }, [user]);
+ 
+
+    const userRef = useRef<User | null>(null);
 
     const login = (user: User) => {
         setUser(user);
+        userRef.current = user;
     };
 
     const logout = () => {
         setUser(null);
+        userRef.current = null;
     };
 
     return (
-        <UserContext.Provider value={{ user: userRef.current, userRef, login, logout }}>
+        <UserContext.Provider value={{ user, setUser, userRef, login, logout, show, setShow}}>
             {children}
         </UserContext.Provider>
     );
